logs/requirements.txt
PIKA==1.2.0
PYTHON-DOTENV==0.15.0